requirements.txt
amqp==5.2.0
appnope==0.1.4
asttokens==3.0.1
attrs==23.2.0
billiard==4.2.4
bump2version==1.0.1
celery==5.6.2
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
dnspython==2.4.2
eventlet==0.34.2
exceptiongroup==1.3.1
executing==2.2.1
fiona==1.9.5
geographiclib==2.0
geopandas==0.14.1
geopy==2.4.1
gmqtt==0.7.0
graphviz==0.21
greenlet==3.0.3
haversine==2.8.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.6.2
librt==0.8.1
matplotlib-inline==0.2.1
mypy==1.19.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.3
orsim==1.2.1
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.4
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pika==1.3.2
platformdirs==4.4.0
pluggy==1.3.0
ply==3.11
polyline==2.0.1
prompt-toolkit==3.0.43
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydot==4.0.1
Pygments==2.19.2
Pyomo==6.7.0
pyparsing==3.3.2
pyproj==3.6.1
pytest==7.4.4
python-dateutil==2.8.2
python-statemachine==3.0.0
pytz==2023.3.post1
pyzmq==27.1.0
requests==2.31.0
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
websockets==12.0
zipp==3.17.0
